requirements.txt
requests
pandas
sqlalchemy
psycopg
matplotlib